Reduce casts between uint and int in PE writer

+++ src/Compilers/Core/Portable/CodeGen/ILBuilder.cs
@@ -877,7 +877,7 @@ namespace Microsoft.CodeAnalysis.CodeGen
                     for (int i = blockFirstMarker; i <= blockLastMarker; i++)
                     {
                         int blockOffset = _allocatedILMarkers[i].BlockOffset;
-                        int absoluteOffset = (int)writer.Position + blockOffset;
+                        int absoluteOffset = writer.Position + blockOffset;
                         _allocatedILMarkers[i] = new ILMarker() { BlockOffset = blockOffset, AbsoluteOffset = absoluteOffset };
                     }
                 }
